+++ src/core/tsi/alts/zero_copy_frame_protector/alts_zero_copy_grpc_protector.cc
@@ -217,8 +217,8 @@ static tsi_result alts_zero_copy_grpc_protector_unprotect(
   }
   if (min_progress_size != nullptr) {
     if (protector->parsed_frame_size > kZeroCopyFrameLengthFieldSize) {
-      *min_progress_size = static_cast<int>(protector->parsed_frame_size -
-                                            protector->protected_sb.length);
+      *min_progress_size =
+          protector->parsed_frame_size - protector->protected_sb.length;
     } else {
       *min_progress_size = 1;
     }
